+++ packages/mangrove.js/CHANGELOG.md
@@ -1,11 +1,14 @@
 # Next version
 
+# 0.2.0 (February 2022)
+
 - New `Market` options:
   - `desiredPrice`: allows one to specify a price point of interest. This will cause the cache to initially load all offers with this price or better.
   - `desiredVolume`: allows one to specify a volume of interest. This will cause the cache to initially load at least this volume (if available). The option uses the same specification as for `estimateVolume`: `desiredVolume: { given: 1, what: "base", to: "buy" }` will cause the asks semibook to be initialized with a volume of at least 1 base token.
 - New `Market` subscription: `market.afterBlock(n,callback)` will trigger `callback` after the market events for block `n` have been processed. If the block has already been processed, `callback` will be triggered at the next event loop.
 - Improve logging: add file logging, allow applications using the package to configure logging using local `config` file.
 - add support for keystore file (json wallet) (`Mangrove.connect(jsonWallet:{path 'path/to/file.json', password: <wallet password>})`)
+- New `partialFill` flag in `OrderResult`: This flag will be true if the order was only partially filled.
 
 # 0.1.0 (January 2022)